Fix sink for low iterations
The signal got stuck in a sink moving back and forth between two
neighbouring positions for a while at the start of every simulation.

As a result of using unsigned ints subtracting something from iter
for low values of iter yeilds a large positive value. Explicitly
casting to int prevents this and gives the small negative result
expected and prevents the signal from becoming stuck.